+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+- Too long flush duration (#2370)
 - Fix issue with invalid profiles uploading (#2358 and #2359)
 
 ## 7.30.0